REFACTOR: Moved routes file into `constants` folder
