+++ app/auth/login/_components/buttons/social-auth-buttons.tsx
@@ -3,7 +3,7 @@
 import { BetterAuthActionButton } from "@/components/auth/buttons/better-auth-action-button";
 import { Button } from "@/components/ui/button";
 import { authClient } from "@/lib/auth/auth-client";
-import { ROUTES } from "@/lib/routes";
+import { ROUTES } from "@/constants/routes";
 import {
   SUPPORTED_OAUTH_PROVIDER_DETAILS,
   SUPPORTED_OAUTH_PROVIDERS,
