+++ app/auth/login/_components/tabs/sign-in-tab.tsx
@@ -15,7 +15,7 @@ import { PasswordInput } from "@/components/ui/password-input";
 import { Button } from "@/components/ui/button";
 import { LoadingSwap } from "@/components/ui/loading-swap";
 import { authClient } from "@/lib/auth/auth-client";
-import { ROUTES } from "@/lib/routes";
+import { ROUTES } from "@/constants/routes";
 import { toast } from "sonner";
 import { useRouter } from "next/navigation";
 import { PasskeyButton } from "../buttons/passkey-button";
