+++ app/auth/login/_components/tabs/sign-up-tab.tsx
@@ -15,7 +15,7 @@ import { PasswordInput } from "@/components/ui/password-input";
 import { Button } from "@/components/ui/button";
 import { LoadingSwap } from "@/components/ui/loading-swap";
 import { authClient } from "@/lib/auth/auth-client";
-import { ROUTES } from "@/lib/routes";
+import { ROUTES } from "@/constants/routes";
 import { toast } from "sonner";
 import { UserPlus } from "lucide-react";
 import { signUpSchema, SignUpForm } from "@/schemas/sign-up";
