+++ app/auth/reset-password/_components/reset-password-form.tsx
@@ -13,7 +13,7 @@ import {
 import { Button } from "@/components/ui/button";
 import { LoadingSwap } from "@/components/ui/loading-swap";
 import { authClient } from "@/lib/auth/auth-client";
-import { ROUTES } from "@/lib/routes";
+import { ROUTES } from "@/constants/routes";
 import { toast } from "sonner";
 import { PasswordInput } from "@/components/ui/password-input";
 import {
