+++ components/assistant/assistant-options.tsx
@@ -7,7 +7,7 @@ import { RenameDialog } from "@/components/shared/rename-dialog";
 import { ResponsiveMenu } from "@/components/shared/responsive-menu";
 import { DeleteConfirmDialog } from "@/components/shared/delete-confirm-dialog";
 import { useRouter } from "next/navigation";
-import { ROUTES } from "@/lib/routes";
+import { ROUTES } from "@/constants/routes";
 import { useCreateChat } from "@/hooks/use-create-chat";
 import { useEntityOptions } from "@/hooks/use-entity-options";
