+++ components/chat/chat-card.tsx
@@ -3,7 +3,7 @@
 import { Card } from "@/components/ui/card";
 import { MessageSquare } from "lucide-react";
 import { useRouter } from "next/navigation";
-import { ROUTES } from "@/lib/routes";
+import { ROUTES } from "@/constants/routes";
 import type { Chat } from "@/types/chat";
 import { ChatOptions } from "./chat-options";
