+++ components/prompt/prompt-card.tsx
@@ -3,7 +3,7 @@
 import { Card } from "@/components/ui/card";
 import { Command } from "lucide-react";
 import { useRouter } from "next/navigation";
-import { ROUTES } from "@/lib/routes";
+import { ROUTES } from "@/constants/routes";
 import type { Prompt } from "@/types/prompt";
 import { PromptOptions } from "./prompt-options";
 import { Badge } from "@/components/ui/badge";
